backend/requirements.txt
aiofiles
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.0.1
cffi==1.17.1
click==8.1.8
colorama==0.4.6
confluent-kafka==2.8.2
cryptography==39.0.2
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.11
greenlet==3.1.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.5
kafka-python==2.0.6
MarkupSafe==3.0.2
numpy==2.2.3
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pluggy==1.5.0
psutil==5.9.4
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyMySQL==1.0.3
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.5
starlette==0.46.0
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.21.0
xlrd >= 2.0.1
openpyxl